requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
pymupdf>=1.24.0
streamlit>=1.38.0
python-dotenv>=1.0.0
